refactor: ControllerScanner의 메서드명 변경
